feat: Add GitHub Actions workflow for tox
This commit introduces a new GitHub Actions workflow to automate the execution of tests using tox.

The workflow is configured to trigger on every push to any branch and runs the test suite against multiple Python versions (3.6, 3.7, 3.8, 3.9, 3.10, 3.11, 3.12) on an Ubuntu environment. This will ensure that all changes are automatically tested, improving the project's stability and code quality.

Signed-off-by: ilias <ekiourk@gmail.com>